LayoutTests/http/tests/subresource-integrity/sri-fetch-expected.txt

CONSOLE MESSAGE: Failed integrity metadata check. http://127.0.0.1:8000/subresource-integrity/resources/resource.txt. Failed integrity metadata check. Content length: 3, Expected content length: 3, Expected metadata: sha256-dKUcPOn/AlUjWIwcHeHNqYXPlvyGiq+2dWOdFcE+24I=
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/subresource-integrity/resources/resource.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://127.0.0.1:8000/subresource-integrity/resources/resource.txt. Failed integrity metadata check. Content length: 3, Expected content length: 3, Expected metadata: sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg== sha384-MgZYnnAzPM/MjhqfOIMfQK5qcFvGZsGLzx4Phd7/A8fHTqqLqXqKo8cNzY3xEPTL
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/subresource-integrity/resources/resource.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://127.0.0.1:8000/subresource-integrity/resources/resource.txt. Failed integrity metadata check. Content length: 3, Expected content length: 3, Expected metadata: sha256-dKUcPOn/AlUjWIwcHeHNqYXPlvyGiq+2dWOdFcE+24I= sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg==
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/subresource-integrity/resources/resource.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://localhost:8000/subresource-integrity/resources/crossorigin-anon-resource.txt. Failed integrity metadata check. Content length: 3, Expected content length: 3, Expected metadata: sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg==
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/subresource-integrity/resources/crossorigin-anon-resource.txt due to access control checks.
CONSOLE MESSAGE: Failed integrity metadata check. http://localhost:8000/subresource-integrity/resources/crossorigin-creds-resource.txt. Failed integrity metadata check. Content length: 3, Expected content length: 3, Expected metadata: sha512-oUceBRNxPxnY60g/VtPCj2syT4wo4EZh2CgYdWy9veW8+OsReTXoh7dizMGZafvx9+QhMS39L/gIkxnPIn41Zg==
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/subresource-integrity/resources/crossorigin-creds-resource.txt due to access control checks.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/subresource-integrity/resources/crossorigin-ineligible-resource.txt due to access control checks.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/subresource-integrity/resources/crossorigin-ineligible-resource.txt due to access control checks.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://localhost:8000/subresource-integrity/resources/crossorigin-ineligible-resource.txt due to access control checks.

PASS Empty string integrity
PASS SHA-256 integrity
PASS SHA-384 integrity
PASS SHA-512 integrity
PASS Invalid integrity
PASS Unknown integrity
PASS Multiple integrities: valid stronger than invalid
PASS Multiple integrities: invalid stronger than valid
PASS Multiple integrities: invalid as strong as valid
PASS Multiple integrities: both are valid
PASS Multiple integrities: both are invalid
PASS Anonymous CORS empty integrity
PASS Anonymous CORS SHA-512 integrity
PASS Anonymous CORS invalid integrity
PASS Credential CORS empty integrity
PASS Credential CORS SHA-512 integrity
PASS Credential CORS invalid integrity
PASS Ineligible CORS empty integrity
PASS Ineligible CORS SHA-512 integrity
PASS Ineligible CORS invalid integrity
PASS SHA-256 integrity with 'no-cors' mode
PASS Resource with zero length body
